# api/src/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: SeferBiz API
  version: 1.0.0
  description: Express-based API rebuilt from the existing route map and resource list.
servers:
  - url: https://api.biz.loc:9443
paths:
  /health:
    get:
      summary: Health check
      responses:
        '200':
          description: Service is healthy
  /v1/invoices:
    get:
      summary: List invoices
      responses:
        '200':
          description: Invoice list
    post:
      summary: Create invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvoiceInput'
      responses:
        '201':
          description: Invoice created
  /v1/invoices/{id}:
    get:
      summary: Get invoice by id
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Invoice record
    put:
      summary: Replace invoice fields
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvoiceInput'
      responses:
        '200':
          description: Invoice updated
    patch:
      summary: Update invoice fields
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvoiceInput'
      responses:
        '200':
          description: Invoice updated
    delete:
      summary: Delete invoice and its line items
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '204':
          description: Invoice deleted
  /v1/invoices/{id}/email:
    get:
      summary: Send invoice email
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Email queued
  /v1/invoices/{invoice_id}/lines:
    get:
      summary: List invoice lines
      parameters:
        - $ref: '#/components/parameters/InvoiceIdParam'
      responses:
        '200':
          description: Invoice lines
    post:
      summary: Create invoice line
      parameters:
        - $ref: '#/components/parameters/InvoiceIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvoiceLineInput'
      responses:
        '201':
          description: Invoice line created
  /v1/invoices/{invoice_id}/lines/{id}:
    get:
      summary: Get invoice line by id
      parameters:
        - $ref: '#/components/parameters/InvoiceIdParam'
        - $ref: '#/components/parameters/LineIdParam'
      responses:
        '200':
          description: Invoice line
  /v1/clients:
    get:
      summary: List clients
      responses:
        '200':
          description: Client list
    post:
      summary: Create client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientInput'
      responses:
        '201':
          description: Client created
  /v1/clients/{id}:
    get:
      summary: Get client by id
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Client record
    put:
      summary: Replace client fields
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientInput'
      responses:
        '200':
          description: Client updated
    patch:
      summary: Update client fields
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientInput'
      responses:
        '200':
          description: Client updated
    delete:
      summary: Delete client
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '204':
          description: Client deleted
  /years:
    get:
      summary: List years
      responses:
        '200':
          description: Years ordered newest first
    post:
      summary: Create financial year
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/YearInput'
      responses:
        '201':
          description: Financial year created
  /years/new:
    get:
      summary: Get defaults for a new year
      responses:
        '200':
          description: New year form values
  /years/{id}:
    get:
      summary: Get year totals
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Year and totals
        '404':
          description: Year not found
    put:
      summary: Replace financial year fields
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/YearInput'
      responses:
        '200':
          description: Financial year updated
    patch:
      summary: Update financial year fields
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/YearInput'
      responses:
        '200':
          description: Financial year updated
    delete:
      summary: Delete financial year
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '204':
          description: Financial year deleted
  /years/{id}/edit:
    get:
      summary: Get a year for editing
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Year record
        '404':
          description: Year not found
  /years/{id}/income:
    get:
      summary: Get paid invoice income for a year
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Income report
        '404':
          description: Year not found
  /years/{id}/expenses:
    get:
      summary: Get expenses for a year
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Expense report
        '404':
          description: Year not found
  /invoices/{id}/email:
    get:
      summary: Queue an invoice email
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Email queued
        '404':
          description: Invoice not found
  /invoices/{id}/stripe:
    get:
      summary: Confirm a Stripe invoice payment
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Payment confirmed
        '400':
          description: Invalid Stripe session
        '404':
          description: Invoice not found
  /expenses:
    get:
      summary: List expenses
      responses:
        '200':
          description: Expense list
    post:
      summary: Create expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExpenseInput'
      responses:
        '201':
          description: Expense created
  /expenses/{id}:
    get:
      summary: Get expense by id
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Expense record
        '404':
          description: Expense not found
    put:
      summary: Replace expense fields
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExpenseInput'
      responses:
        '200':
          description: Expense updated
    patch:
      summary: Update expense fields
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExpenseInput'
      responses:
        '200':
          description: Expense updated
    delete:
      summary: Delete expense
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '204':
          description: Expense deleted
  /vendors:
    get:
      summary: List vendors
      responses:
        '200':
          description: Vendor list
    post:
      summary: Create vendor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VendorInput'
      responses:
        '201':
          description: Vendor created
  /vendors/{id}:
    get:
      summary: Get vendor by id
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Vendor record
        '404':
          description: Vendor not found
    put:
      summary: Replace vendor fields
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VendorInput'
      responses:
        '200':
          description: Vendor updated
    patch:
      summary: Update vendor fields
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VendorInput'
      responses:
        '200':
          description: Vendor updated
    delete:
      summary: Delete vendor
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '204':
          description: Vendor deleted
  /charts_controller/trailing_x_months/{number_of_months}:
    get:
      summary: Get trailing monthly invoice, expense, and goal data
      parameters:
        - name: number_of_months
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 120
      responses:
        '200':
          description: Trailing monthly chart series
        '400':
          description: Invalid month count
  /charts_controller/year_invoice_month/{id}:
    get:
      summary: Get monthly paid invoice totals
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Monthly invoice chart series
        '404':
          description: Year not found
  /charts_controller/year_invoice_month_with_goal/{id}:
    get:
      summary: Get monthly invoice totals and goals
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Monthly invoice and goal chart series
        '404':
          description: Year not found
  /charts_controller/year_all_data/{id}:
    get:
      summary: Get invoice, expense, and goal data for a year
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Monthly chart series
        '404':
          description: Year not found
  /charts_controller/year_expense_category/{id}:
    get:
      summary: Get expense totals by category
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Expense category totals
        '404':
          description: Year not found
  /charts_controller/year_expense_month/{id}:
    get:
      summary: Get monthly expense totals
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Monthly expense chart series
        '404':
          description: Year not found
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
    InvoiceIdParam:
      name: invoice_id
      in: path
      required: true
      schema:
        type: integer
    LineIdParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
  schemas:
    InvoiceInput:
      type: object
      required:
        - client_id
      properties:
        client_id:
          type: integer
        status:
          type: string
        date:
          type: string
          format: date
        cost:
          type: number
        total:
          type: number
          description: Accepted as an alias for cost.
        currency:
          type: string
        description:
          type: string
    ClientInput:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        contact:
          type: string
        email_accounting:
          type: string
          format: email
        site_url:
          type: string
          format: uri
        currentrate:
          type: number
        city:
          type: string
        state:
          type: string
    VendorInput:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        category:
          type: string
        notes:
          type: string
    YearInput:
      type: object
      required:
        - year
      properties:
        year:
          type: integer
        taxrate:
          type: number
        goal_year:
          type: number
        goals_months:
          type: array
          minItems: 12
          maxItems: 12
          items:
            type: number
    ExpenseInput:
      type: object
      required:
        - name
        - date
        - cost
      properties:
        name:
          type: string
        vendor_id:
          type: integer
          nullable: true
        date:
          type: string
          format: date
        cost:
          type: number
        account:
          type: string
        notes:
          type: string
    InvoiceLineInput:
      type: object
      required:
        - description
      properties:
        description:
          type: string
        amount:
          type: number
    Invoice:
      type: object
      properties:
        id:
          type: integer
        client_id:
          type: integer
        status:
          type: string
        total:
          type: number
        currency:
          type: string
    Client:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string